+++ src/HearThis/UI/AndroidSyncDialog.cs
@@ -14,55 +14,65 @@ using System.Net.Sockets;
 using System.Text;
 using System.Threading;
 using System.Windows.Forms;
+using L10NSharp;
 using SIL.Reporting;
 using SIL.Windows.Forms.Progress;
 using ZXing;
+using static System.String;
+using static System.Windows.Forms.MessageBoxButtons;
 
 namespace HearThis.UI
 {
 	/// <summary>
-	/// This dialog is responsible to obtain the IP address of the Android we want to sync with, if possible.
-	/// Currently this is done by displaying a QR code representing our own IP address, which the client
-	/// must set using SetOurIpAddress. The android scans this and sends a packet containing its own IP address.
-	/// When this class receives that successfully, it closes (with DialogResult.OK). If instead the user
-	/// closes the dialog, it will return another code, I think Cancel.
-	/// Enhance JT: handle the possibility that the Android does not have a functional scanner.
-	/// Display the Android's IP address on its screen, and tell the user to type it in a new text box (or 4)
-	/// here, then click OK.
-	/// Enhance JT: possibly the QR code could convey a BlueTooth IP address also as another option.
+	/// This dialog is responsible to obtain the IP address of the Android we want to sync with,
+	/// if possible. Currently, this is done by displaying a QR code representing our own IP
+	/// address, which the client must set using <see cref="SetOurIpAddress"/>. The Android device
+	/// scans this and sends a packet containing its own IP address. When this class receives that
+	/// successfully, it closes (with <see cref="DialogResult.OK"/>). If instead the user closes
+	/// the dialog, it will return another code, I think Cancel.
+	/// Enhance JT: handle the possibility that the Android device does not have a functional
+	/// scanner.
+	/// Display the Android device's IP address on its screen, and tell the user to type it in a
+	/// new text box (or 4) here, then click OK.
+	/// Enhance JT: possibly the QR code could alternatively convey a BlueTooth IP address.
 	/// </summary>
-	public partial class AndroidSyncDialog : Form
+	public partial class AndroidSyncDialog : Form, ILocalizable
 	{
 		private UDPListener _listener;
+		private IPAddress _ourIpAddress;
 
 		public event EventHandler<EventArgs> GotSync;
-
-		public LogBox ProgressBox { get; private set; }
-		public AndroidSyncDialog()
-		{
-			InitializeComponent();
-			// This works around a weird behavior of BetterLinkLabel, where the appearance of IsTextSelectable  = false
-			// is achieved by making enabled false. But we want the user to be able to click the link!
-			// Since the purpose of the "Better" label is to handle multi-line and we don't need that, if a link like this
-			// becomes permanent (e.g., a simple link to HTA on Play Store), consider using an ordinary LinkLabel.
-			playStoreLinkLabel.Enabled = true;
-		}
-
+		
 		/// <summary>
 		/// The result, if any, we obtained: the IP address of the Android we should sync with.
 		/// </summary>
 		public static string AndroidIpAddress { get; set; }
 
-		private string _ourIpAddress;
+		public LogBox ProgressBox { get; private set; }
+		public AndroidSyncDialog()
+		{
+			InitializeComponent();
+			// This works around a weird behavior of BetterLinkLabel, where the appearance of IsTextSelectable = false
+			// is achieved by making enabled false. But we want the user to be able to click the link!
+			// Since the purpose of the "Better" label is to handle multi-line and we don't need that, if a link like this
+			// becomes permanent (e.g., a simple link to HTA on Play Store), consider using an ordinary LinkLabel.
+			_playStoreLinkLabel.Enabled = true;
+
+			Program.RegisterLocalizable(this);
+			HandleStringsLocalized();
+		}
 
 		/// <summary>
 		/// Set the IP address (on the wireless network) of this computer.
 		/// This will be displayed as a QR code for the Android to read.
 		/// </summary>
-		/// <param name="content"></param>
-		public void SetOurIpAddress(string content)
+		/// <param name="ipAddress"></param>
+		public void SetOurIpAddress(IPAddress ipAddress)
 		{
-			_ourIpAddress = content;
+			if (ipAddress.AddressFamily != AddressFamily.InterNetwork)
+				throw new ArgumentException("Only IPv4 IP addresses may be used for Android sync.");
+
+			_ourIpAddress = ipAddress;
 			var writer = new BarcodeWriter
 			{
 				Format = BarcodeFormat.QR_CODE,
@@ -72,10 +82,9 @@ namespace HearThis.UI
 					Width = qrBox.Width
 				}
 			};
-			var matrix = writer.Write(content);
+			var matrix = writer.Write(ipAddress.ToString());
 			var qrBitmap = new Bitmap(matrix);
 			qrBox.Image = qrBitmap;
-
 		}
 
 		public void ShowAndroidIpAddress()
@@ -87,12 +96,23 @@ namespace HearThis.UI
 			else if (_ourIpAddress != null)
 			{
 				// We expect it to be on the same network, so the first three groups should be the same
-				int index = _ourIpAddress.LastIndexOf(".", StringComparison.Ordinal);
-				if (index > 0)
+				var bytes = _ourIpAddress.GetAddressBytes(); // e.g., [192, 168, 1, 10]
+				if (bytes.Length == 4)
 				{
-					_ipAddressBox.Text = _ourIpAddress.Substring(0, index + 1) + "???";
+					// REVIEW: Should "???" be localizable?
+					_ipAddressBox.Text = $"{bytes[0]}.{bytes[1]}.{bytes[2]}.???";
 				}
 			}
+		}
+		public void HandleStringsLocalized()
+		{
+			Text = Format(Text, Program.kAndroidAppName);
+			_lblAboutHearThisAndroid.Text = Format(_lblAboutHearThisAndroid.Text,
+				Program.kAndroidAppName, Program.kProduct);
+			_lblSyncInstructions.Text = Format(_lblSyncInstructions.Text,
+				Program.kAndroidAppName);
+			_playStoreLinkLabel.Text = Format(_playStoreLinkLabel.Text, Program.kAndroidAppName);
+			_lblAltIP.Text = Format(_lblAltIP.Text, Program.kAndroidAppName, _syncButton.Text);
 		}
 
 		protected override void OnShown(EventArgs e)
@@ -101,9 +121,10 @@ namespace HearThis.UI
 			_listener = new UDPListener();
 			_listener.NewMessageReceived += (sender, args) =>
 			{
-				AndroidIpAddress = Encoding.UTF8.GetString(args.data);
-				Invoke(new Action(HandleGotIpAddress));
+				AndroidIpAddress = Encoding.UTF8.GetString(args.Data);
+				Invoke(new Action(HandleGotIPAddress));
 			};
+			
 			int index = _ipAddressBox.Text.LastIndexOf(".", StringComparison.Ordinal);
 			_ipAddressBox.SelectionStart = index + 1;
 			_ipAddressBox.SelectionLength = 3;
@@ -129,18 +150,19 @@ namespace HearThis.UI
 		/// Hides the QR code, which has served its purpose, and shows a LogBox to report
 		/// progress of the sync.
 		/// </summary>
-		private void HandleGotIpAddress()
+		private void HandleGotIPAddress()
 		{
 			qrBox.Hide();
 			ProgressBox = new LogBox() ;
 			int progressMargin = 10;
 			ProgressBox.Location = new Point(progressMargin, qrBox.Top);
-			ProgressBox.Size = new Size(this.DisplayRectangle.Width - progressMargin * 2, okButton.Top - ProgressBox.Top - progressMargin);
+			ProgressBox.Size = new Size(DisplayRectangle.Width - progressMargin * 2,
+				_btnClose.Top - ProgressBox.Top - progressMargin);
 			ProgressBox.Anchor = AnchorStyles.Top | AnchorStyles.Left | AnchorStyles.Bottom | AnchorStyles.Right;
 			Controls.Add(ProgressBox);
 			_ipAddressBox.Hide();
 			_syncButton.Hide();
-			_altIpLabel.Hide();
+			_lblAltIP.Hide();
 			ProgressBox.Show();
 			GotSync?.Invoke(this, EventArgs.Empty);
 		}
@@ -149,10 +171,10 @@ namespace HearThis.UI
 		/// Helper class to listen for a single packet from the Android. Construct an instance to start
 		/// listening (on another thread); hook NewMessageReceived to receive a single packet.
 		/// </summary>
-		class UDPListener
+		private class UDPListener
 		{
-			private int _portToListen = 11007; // must match HearThisAndroid SyncActivity.desktopPort
-			public event EventHandler<MyMessageArgs> NewMessageReceived;
+			private const int kPortToListen = 11007; // must match HearThisAndroid SyncActivity.desktopPort
+			public event EventHandler<UDPListenerMessageArgs> NewMessageReceived;
 			UdpClient _listener;
 			private bool _listening;
 
@@ -171,23 +193,24 @@ namespace HearThis.UI
 			{
 				try
 				{
-					_listener = new UdpClient(_portToListen);
+					_listener = new UdpClient(kPortToListen);
 				}
 				catch (SocketException)
 				{
-					//do nothing
+					// Do nothing
 				}
 
 				if (_listener != null)
 				{
-					IPEndPoint groupEP = new IPEndPoint(IPAddress.Any, 0);
+					var groupEndPt = new IPEndPoint(IPAddress.Any, 0);
 
 					try
 					{
-						byte[] bytes = _listener.Receive(ref groupEP); // waits for packet from Android.
+						// Wait for packet from Android.
+						byte[] bytes = _listener.Receive(ref groupEndPt);
 
 						//raise event
-						NewMessageReceived(this, new MyMessageArgs(bytes));
+						NewMessageReceived?.Invoke(this, new UDPListenerMessageArgs(bytes));
 						_listening = false;
 						_listener.Close();
 					}
@@ -209,79 +232,82 @@ namespace HearThis.UI
 		}
 
 		/// <summary>
-		/// Helper class to hold the data we got from the Android, for the NewMessageReceived event of UDPListener
+		/// Helper class to hold the data we got from the Android, for the NewMessageReceived
+		/// event of UDPListener
 		/// </summary>
-		class MyMessageArgs : EventArgs
+		private class UDPListenerMessageArgs : EventArgs
 		{
-			public byte[] data { get; set; }
+			public byte[] Data { get; }
 
-			public MyMessageArgs(byte[] newData)
+			public UDPListenerMessageArgs(byte[] newData)
 			{
-				data = newData;
+				Data = newData;
 			}
 		}
 
 		private void _syncButton_Click(object sender, EventArgs e)
 		{
 			AndroidIpAddress = _ipAddressBox.Text;
+
+			string GetProblemCaption() =>
+				Format(LocalizationManager.GetString("AndroidSynchronization.SyncProblemCaption",
+					"{0} Sync Problem",
+					"MessageBox caption; Param is \"HearThis\" (product name)"),
+					Program.kProduct);
+			
 			if (AndroidIpAddress.Contains("?"))
 			{
 				MessageBox.Show(
-					"You need to replace the three question marks in the box to the left with the number shown on the device in order to sync manually.",
-					"HearThis Sync Problem");
+					LocalizationManager.GetString("AndroidSynchronization.ReplaceQuestionMarks",
+					"You need to replace the three question marks in the address box with the number shown on the Android device in order to sync manually."),
+					GetProblemCaption());
 				return;
 			}
 
-			if (!ValidateIpAddress())
+			if (!IPAddress.TryParse(AndroidIpAddress, out var ipAddress) ||
+			    ipAddress.AddressFamily != AddressFamily.InterNetwork)
 			{
 				MessageBox.Show(
-					"The value in the address box does not appear to be a valid device address.",
-					"HearThis Sync Problem");
+					Format(LocalizationManager.GetString("AndroidSynchronization.InvalidIpAddress",
+						"The value in the address box does not appear to be a valid ({0}) " +
+						"device address.", "\"IPv4\""), "IPv4"),
+					GetProblemCaption());
 				return;
 			}
 
-			if (!IsPlausibleIpAddress())
+			if (!IsPlausibleIPAddress(ipAddress))
 			{
-				if (MessageBox.Show(
-					"The device address you entered appears to be on a different local network. This usually won't work. Do you want to try anyway?",
-					"Warning", MessageBoxButtons.YesNo) == DialogResult.No)
+				var result = MessageBox.Show(
+					LocalizationManager.GetString("AndroidSynchronization.DifferentNetwork",
+						"The device address you entered appears to be on a different local network. This usually " +
+						"won't work. Do you want to try anyway?"),
+					LocalizationManager.GetString("AndroidSynchronization.WarningCaption",
+						"Warning"), YesNo);
+				if (result == DialogResult.No)
 				{
 					return;
 				}
 			}
 			_listener.StopListener();
-			HandleGotIpAddress();
+			HandleGotIPAddress();
 		}
 
-		private bool ValidateIpAddress()
+		private bool IsPlausibleIPAddress(IPAddress ipAddress)
 		{
-			var parts = AndroidIpAddress.Split('.');
-			if (parts.Length != 4)
+			var ourBytes = _ourIpAddress.GetAddressBytes();
+			var otherBytes = ipAddress.GetAddressBytes();
+
+			// Only compare IPv4 addresses (caller should have already validated)
+			if (ourBytes.Length != 4 || otherBytes.Length != 4)
 				return false;
-			foreach (var part in parts)
+
+			// Compare first 3 bytes (prefix)
+			for (int i = 0; i < 3; i++)
 			{
-				if (!int.TryParse(part, out var val) || val < 0 || val > 255)
+				if (ourBytes[i] != otherBytes[i])
 					return false;
 			}
 			return true;
-		}
-
-		private bool IsPlausibleIpAddress()
-		{
-			return getIpPrefix(_ourIpAddress) == getIpPrefix(AndroidIpAddress);
-		}
-
-		private string getIpPrefix(string input)
-		{
-			int index = input.LastIndexOf('.');
-			if (index < 0)
-				return "";
-			return input.Substring(0, index + 1);
-		}
-
-		private void okButton_Click(object sender, EventArgs e)
-		{
-			Close();
 		}
 	}
 }