+++ DistFiles/ReleaseNotes.md
@@ -39,7 +39,8 @@ the recorded files to that format, if necessary.
 ## _VERSION_ (_DATE_)
 - Improvements related to opening a clip for editing in an external program.
 - Updated localizations.
-- Improved connectivity with HearThisAndroid.
+- Improved connectivity with HearThis Android.
+- Made UI strings related to HearThis Android localizable.
 
 ## 3.5.3 (March 2025)
 - Fixed bug affecting Glyssenscript projects when extra clips are present.